Improved the look of the game and added new functionality
Players can now play continuously without having to rerun the program. 
Players can also go back and choose another level if they are not satisfied with the level they picked. For example, if a user picks "Easy" as the level, but suddenly they have a change of heart, the game will ask "Are you sure you want to play this level?". If the user selects "No", they will be able to choose a different level. If the user selects "Yes", they will enter the game.